Move command log structs into tedge_api
