--- crates/common/logged_command/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "logged_command"
-version = { workspace = true }
-authors = { workspace = true }
-edition = { workspace = true }
-rust-version = { workspace = true }
-license = { workspace = true }
-homepage = { workspace = true }
-repository = { workspace = true }
-
-[dependencies]
-camino = { workspace = true }
-log = { workspace = true }
-nix = { workspace = true }
-shell-words = { workspace = true }
-tedge_api = { workspace = true }
-time = { workspace = true, features = ["formatting"] }
-tokio = { workspace = true, features = [
-    "fs",
-    "io-util",
-    "macros",
-    "process",
-    "rt",
-] }
-
-[dev-dependencies]
-anyhow = { workspace = true }
-tedge_test_utils = { workspace = true }
-tokio = { workspace = true, features = ["time"] }
-
-[lints]
-workspace = true
