--- crates/common/logged_command/src/lib.rs
@@ -1,5 +1,0 @@
-mod logged_command;
-
-pub use crate::logged_command::CommandLog;
-pub use crate::logged_command::LoggedCommand;
-pub use crate::logged_command::LoggingChild;
