+++ crates/common/tedge_utils/src/signals.rs
@@ -1,3 +1,4 @@
+use nix::unistd::Pid;
 use std::io;
 
 #[cfg(not(windows))]
@@ -15,4 +16,21 @@ pub async fn interrupt() -> io::Result<()> {
 #[cfg(windows)]
 pub async fn interrupt() -> io::Result<()> {
     tokio::signal::ctrl_c().await
+}
+
+pub enum Signal {
+    SIGTERM,
+    SIGKILL,
+}
+
+pub fn terminate_process(pid: u32, signal_type: Signal) {
+    let pid: Pid = nix::unistd::Pid::from_raw(pid as nix::libc::pid_t);
+    match signal_type {
+        Signal::SIGTERM => {
+            let _ = nix::sys::signal::kill(pid, nix::sys::signal::SIGTERM);
+        }
+        Signal::SIGKILL => {
+            let _ = nix::sys::signal::kill(pid, nix::sys::signal::SIGKILL);
+        }
+    }
 }
