+++ crates/core/plugin_sm/src/plugin.rs
@@ -1,8 +1,6 @@
 use async_trait::async_trait;
 use csv::ReaderBuilder;
 use download::Downloader;
-use logged_command::CommandLog;
-use logged_command::LoggedCommand;
 use regex::Regex;
 use reqwest::Identity;
 use serde::Deserialize;
@@ -10,7 +8,14 @@ use std::error::Error;
 use std::path::Path;
 use std::path::PathBuf;
 use std::process::Output;
-use tedge_api::*;
+use tedge_api::CommandLog;
+use tedge_api::DownloadInfo;
+use tedge_api::LoggedCommand;
+use tedge_api::SoftwareError;
+use tedge_api::SoftwareModule;
+use tedge_api::SoftwareModuleUpdate;
+use tedge_api::SoftwareType;
+use tedge_api::DEFAULT;
 use tedge_config::SudoCommandBuilder;
 use tokio::io::AsyncWriteExt;
 use tracing::error;
