+++ crates/core/tedge_agent/src/software_manager/actor.rs
@@ -6,7 +6,6 @@ use crate::state_repository::error::StateError;
 use crate::state_repository::state::AgentStateRepository;
 use anyhow::anyhow;
 use async_trait::async_trait;
-use logged_command::CommandLog;
 use plugin_sm::plugin_manager::ExternalPlugins;
 use plugin_sm::plugin_manager::Plugins;
 use serde::Deserialize;
@@ -30,6 +29,7 @@ use tedge_api::mqtt_topics::OperationType;
 use tedge_api::workflow::GenericCommandData;
 use tedge_api::workflow::GenericCommandMetadata;
 use tedge_api::workflow::GenericCommandState;
+use tedge_api::CommandLog;
 use tedge_api::Jsonify;
 use tedge_api::SoftwareType;
 use tedge_config::TEdgeConfigError;
