+++ crates/core/tedge_agent/src/tedge_operation_converter/actor.rs
@@ -4,7 +4,6 @@ use async_trait::async_trait;
 use camino::Utf8PathBuf;
 use log::error;
 use log::info;
-use logged_command::CommandLog;
 use std::process::Output;
 use std::time::Duration;
 use tedge_actors::fan_in_message_type;
@@ -29,8 +28,10 @@ use tedge_api::workflow::GenericCommandMetadata;
 use tedge_api::workflow::GenericCommandState;
 use tedge_api::workflow::GenericStateUpdate;
 use tedge_api::workflow::OperationAction;
+use tedge_api::workflow::OperationName;
 use tedge_api::workflow::WorkflowExecutionError;
 use tedge_api::workflow::WorkflowSupervisor;
+use tedge_api::CommandLog;
 use tedge_mqtt_ext::MqttMessage;
 use tedge_mqtt_ext::QoS;
 use tedge_script_ext::Execute;
