+++ crates/core/tedge_api/Cargo.toml
@@ -29,6 +29,7 @@ time = { workspace = true, features = [
     "serde",
     "serde-well-known",
 ] }
+tokio = { workspace = true, features = ["fs", "process"] }
 
 [dev-dependencies]
 anyhow = { workspace = true }
@@ -36,6 +37,7 @@ assert_matches = { workspace = true }
 clock = { workspace = true }
 maplit = { workspace = true }
 mockall = { workspace = true }
+tedge_test_utils = { workspace = true }
 tempfile = { workspace = true }
 test-case = { workspace = true }
 time = { workspace = true, features = ["macros"] }
