+++ crates/core/tedge_api/src/workflow/log/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod command_log;
+pub(crate) mod logged_command;
